Import Mini vMac 0.1.7
version 0.1.7 released. It fixes a bug noticed when a program doing
floating point computations gave wrong answers. The problem was more
general though. (Specifically, that moving to and from the status
register shouldn't set the condition codes like most other move
operations. It happened because the CPU emulation of Mini vMac descended
from a disassembler, where the distinction doesn't matter.)

This version has a new feature where disk image files named 'disk1.dsk',
'disk2.dsk', and 'disk3.dsk' will, if present in the folder containing
Mini vMac, be opened when Mini vMac is launched. In the Mac versions,
these files can be aliases to the real image disk image files. Also the
'vMac.ROM' file can now be an alias file in the Mac version.

The big new feature is an 'autoquit' capability. It assists in
'wrapping' an application for the Mac Plus, so that you can use it more
as you would an ordinary application for your machine. Actually this was
already in Mini vMac, but it wasn't publicly accessible.

The format of the documentation has changed slightly, and the 'About'
page now starts with a nice quote.