Import Mini vMac 1.0.0
version 1.0.0 released. This is 1.0 not because there is any great
change from 0.1.7, but because there isn't. It indicates this is now a
stable branch, with no further changes except for bug fixes. I might
some time fork off an unstable branch, to try out a few ideas, but the
1.0.x branch will continue to be maintained.

This release adds a small kludge to the carbon version for Mac OS X, so
that Commmand-Q and Command-H will be passed on to the emulator, rather
than intercepted by the operating system. This is accomplished by adding
fake menu items. I'd be interested if any one knows of a cleaner way to
do this.

The Windows version no longer registers itself in the Windows registry,
to avoid the possibility of causing problems for other programs that
want to use the '.dsk' and '.rom' extensions. The versions for the X
Window System and for Mac OS 9 and earlier are nearly unchanged.

This biggest change is to adopt the build method from Gryphel, which
might make it easier to build Mini vMac from the source code. It also
supports more development environments. (By the way, people who are
interested in Mini vMac as an example of how to write an emulator might
also be interested in Gryphel as an illustration of how to implement a
graphical user interface. Though it might be more interesting when it is
more finished.)