Import Mini vMac 2.5.0
version 2.5.0 officially released. Some highlights are emulation of the
extended Parameter RAM, assorted fixes, and more speed in the Macintosh
version.

Mini vMac now emulates the extended Parameter RAM of the Mac Plus,
thanks to code contributed by Philip Cummins. Previously only the 20
byte PRAM of the original Macintosh 128 was emulated. The larger 256
byte PRAM stores additional information such as the time zone, latitude,
longitude, and the alert sound. So now the Map and Sound control panels
will work. This information is not saved on quit. The Macintosh version
will initialize the time zone, latitude, and longitude from the real
machine's information. The Windows version only initializes the time
zone. I have some code that can find the time zone for the X version,
but I'm not sure how portable it is, so it is disabled.

The emulation of the SCC chip is improved a bit, so that booting with
AppleTalk turned on won't hang, and printing to the imagewriter will
successfully do nothing. The SCC code from the Windows version of vMac
was used as a starting point, with permission from Weston Pawlowski.
(But unlike vMac for Windows, Mini vMac only tries to emulate serial
ports that aren't connected to anything.) Also, thanks to Philip Cummins
for pointing out the Zilog documentation for the SCC chip.

A small change to the VIA timer 2 code seems to stop the tendency to
hang playing sounds in System 7. It seems to work ok, even though it
doesn't fix what is probably the real problem, that Mini vMac only
resolves time to about a millisecond, while the real VIA timer has much
higher resolution.

I've added a number of programs that now work because of this change to
the currently available Software page, along with many other additions.
Over 40 have been added, bringing the total to over 200.

The PowerPC assembly language is tweaked a bit more, making the
Macintosh version about 20 to 40 percent faster.

The Restart command will now work in all system versions. It seems that,
contrary to the impression given by the documentation, executing the
68000 RESET instruction doesn't change the memory overlay state. Or at
least, empirically, that seems to work better.

Some subtle bugs involving word access to address registers have been
fixed. The main known effect is that the game BachMan sounds different.

Bad accesses to disk images could previously cause data to be written
past the end of the memory allocated for the emulated RAM, which in
theory could cause Mini vMac to crash or worse. This has been fixed.

In the Macintosh version, the F1 and F2 keys are now mapped to emulated
option and command keys, to match the Windows and X versions. This
provides a way to type some key combinations that are intercepted by the
OS of the real machine. For example F2-Shift-3 will take a screen shot
in the emulated machine, whereas Command-Shift-3 takes a screen shot of
the real Macintosh at the same time.

The interrupt button emulation is now more realistic. One consequence is
that the interrupt button can be masked, if the processor's priority
level is high enough.

Previously, in the Windows version, double clicking on a disk image in
the open dialog would tend not to work, because the mouse button would
still be down upon returning to Mini vMac, which prevents mounting.
Mouse clicks are now handled differently, which solves this problem.

Previously, in the X version, quitting Mini vMac could be delayed if the
window manager moves the window to the back when it is being closed
(such as in Red Hat Linux 9). This is fixed.

Mini vMac now has a mechanism for reporting reaching a part of the
emulation that hasn't been implemented, or at least never tested.

A version compiled for Macintosh 680x0 is now included. So that it can
run inside Mini vMac, this version emulates a Macintosh Plus with only
2M of memory instead of 4M. This is more amusing than useful. Since it
is so very slow, I've not tested it much.