Import Mini vMac 3.5.6
Today's Mini vMac 3.5.6 beta fixes an issue in the Cocoa port and fixes
a design flaw in "Abnormal Situation" reports.

The fix in Mini vMac 3.5.5, Cocoa port, for incorrect cursor position
while dragging the Mini vMac window caused another problem. If you
switched to Mini vMac using command-tab, the mouse position wasn't
updated until the mouse is moved. This is fixed by fetching the current
mouse position when Mini vMac is activated, rather than only rely on
mouse movement events. (Before 3.5.5, the Cocoa port would frequently
fetch the mouse position, which was giving incorrect results during
dragging the Mini vMac window.)

If Mini vMac reports an "Abnormal Situation", it now also displays a 4
digit hexadecimal number. The original idea was that if something
unexpected happens, you should first figure out how to reproduce the
problem. Then the same thing can be done on a copy of Mini vMac with
debugging stuff enabled. But this idea doesn't work out so well if the
problem can't be reproduced, and seems to happen randomly and rarely.
Which is what has happened. Over a few weeks I saw a couple Abnormal
Situation reports in the beta that I couldn't reproduce. Mini vMac
worked fine afterwards, so I suspect that the report was not needed, it
is just a situation that is rare but valid. I just need to find out
which report, out of couple hundred, to remove. So I enabled debug
logging and continued to use Mini vMac for weeks. And so far it has
worked perfectly, unfortunately. Maybe I just haven't used it enough.

So now in today's version the regular non debug Mini vMac will display a
number when reporting an Abnormal Situation, which doesn't take much
additional code. So if anyone runs across this issue, you can tell me
this 4 digit hexadecimal number, and I will know exactly which report to
remove.

(By the way, an explanation for all these "Abnormal Situation" reports:
When trying to emulate some aspect of Macintosh hardware, my current
strategy is not to try to figure it all out and then implement the
entire emulation as best I can. Because if I did that, inevitably there
would be a mistake somewhere, and when that causes some software to
malfunction, I won't have any clue of where to look for the mistake.
Instead I would start with a place holder Abnormal Situation report.
Then I find some Macintosh software that reaches that report, and look
at that area of its code, and try to figure out what it is trying to do.
Then I try to implement just enough emulation for it to work correctly,
and give an Abnormal Situation report for code that does something else.
If the Macintosh software still doesn't work, the problem is probably in
the code I just wrote. If it does work, then I next try find some other
Macintosh software that hits one of the new Abnormal Situation reports.
And iterate. The result is code that is usually correct for what it
does, and will tell you exactly where to look when it reaches something
that hasn't been implemented yet.)