Import Mini vMac 3.5.8
Today's Mini vMac 3.5.8 updates the stable version to fix a problem on
PowerPC OS X, and also fixes an issue affecting the Variation Service.
Mini vMac 3.5.8 on platforms other than PowerPC OS X ('mach'), and
x86-32 OS X ('imch'), should be identical to Mini vMac 3.5.7, except for
the version string and modification date.

It was reported that "Mini vMac 3.5.7 wont run on PPC G3 systems". It
turns out that the GCC flag "-mmacosx-version-min" should be specified
for all files compiled, not just the platform dependent code. It affects
things like the required CPU. Making this change happens to have no
effect on Mini vMac for x86-64 OS X, there is some effect for x86-32 OS
X, and the largest effect is for PowerPC.