Import Mini vMac 180325
Today's Mini vMac Development source snapshot adds a new build option
"-iid 1" to enable a new "Insert Ith Disk Image" feature. When this
feature is enabled, if the Control key is held down and a number key
from '1' to '9' is pressed, then Mini vMac will try to mount a disk
image named from "disk1.dsk" to "disk9.dsk" in the folder containing the
application.

This is the same series of disk image names that are automatically
mounted when Mini vMac is launched. But it stops at the first image not
found. So if you leave a name unused, then you can use a control-number
key to mount disks after launch. Or, you can use a control-number key to
remount a disk that was automatically mounted at launch and then later
ejected.

One example use is if you have one copy of Mini vMac running a
development environment (such as MPW) that is used to compile a program
to a disk image. The disk image is then unmounted and mounted on another
copy of Mini vMac running a testing environment. If the compiled program
crashes badly, the development environment is not disturbed. The
control-number key feature makes it easier to move the disk image back
and forth between the two copies of Mini vMac.

Also, the version numbering scheme has changed as of this snapshot.
Instead of "3.6.0", it is "36.00". The distinction between major and
minor version numbers has not been useful. Instead there is now a single
branch number. This allows more bug fix releases (up to 99 instead of up
to 9) without taking more space in the version string.