Import Mini vMac 180401
Today's Mini vMac Development source snapshot adds a new 'P' command to
the Mini vMac Control Mode, which copies to the clipboard (of the real
computer, not the emulated computer) a string containing the options
describing this variation of Mini vMac.

The string generated by the 'P' command starts with a new option, "-br",
which is for future use and currently has no effect. It tells the build
system that the requested options were from a specific branch (version)
of Mini vMac. So "-br 36" means from branch 36. If the default settings
have changed from since that older version, then the build system will
use defaults of the old version, rather than the current defaults. So if
you use the 'P' command on some variation of the current alpha version
of Mini vMac, and pass it to the build system of some future version,
the newly generated variation will behave as close as possible to the
earlier variation.

There are three new build options that may have different defaults in
some future version: "-eci", "-ecr", and "-eck". They can be used to
disable the 'I', 'R', and 'K' commands of the Mini vMac control mode,
which are use to operate the emulated Interrupt Button, emulated Reset
Button, and emulated Control Key. In the future they may be disabled by
default. In the original Macintosh, the Interrupt and Reset buttons were
disabled by default. You needed to install the "Programmer's Switch" to
use them. The original Macintosh did not have a control key. And, using
the 'K' command accidently would be confusing.