Import Mini vMac 180413
Today's Mini vMac Development source snapshot adds a number of features
to the build system. The motivation is that the Variations Service page
gets unwieldy as more and more options are added. So I'm thinking of
instead having a Basic Variations Service, which only has the most
popular options, and a new Advanced Variations Service, which allows
pasting in options as text. (With the new 'P' command of the Mini vMac
Control Mode to get the options of an existing variation, you would be
able to easily update a variation.)

If the new option "@" is used, no developer options are permitted to the
right of it. This will allow the Variations Service to allow the user to
specify options as text without the chance of breaking the service. (For
example, the '-e' option to select the the development environment would
not work well.)

The new option "-ef 1" causes the build system to write error messages
to a text file, instead of displaying an alert dialog. This will allow
the Variations Service to allow the user to specify options as text, and
get any error messages, such as about bad syntax, back to the web page
shown to the user, rather than breaking the service.

If the new option "!" is used, options to the right of it will override
options to the left of it. Normally it would be an error for the same
option to be used twice. It is still an error for an option to be used
twice on the same side of '!'.

The new option "-cte 1" causes the build system to generate source code
that causes the compiler to generate an error. This is for testing the
variation service.

If the new option "-bte" is used, the build system will treat it as an
error. This is for testing the variation service.