Import Mini vMac 180429
Today's Mini vMac Development source snapshot adds the new build system
options "-hcr, -hcg, and -hcb", to select the initial value of red,
green, and blue components of the highlight color in the Parameter RAM.
This was requested by Karl.

Also in today's snapshot, the error message generated by the build
system always include the name of the option which had the problem,
making error messages from the Advanced Variations Service more useful.
(If you use the build system application directly, not telling it to
save error messages to a file and quit, it would usually highlight the
text of where the problem was.)