Import Mini vMac 180610
Today's Mini vMac Development source snapshot fixes a reported bug where
the Control Mode didn't work on ports other than OS X. This was broken
when the "-km" option was added.

Also, I've put in a work around for the Path Randomization misfeature
added in macOS Sierra (10.12). If an application that Apple thinks is
trustworthy is bundled together with malicious library code in the same
folder, the malicious code may be run by the application. The silly way
that Apple has chosen to prevent this is to in effect move the
application somewhere else before running it, so it can't find the
library code. It can't find anything else in the application's folder
either, which is a problem for Mini vMac, which looks for things in its
folder, such as the ROM image file. So now the Cocoa port of Mini vMac
will try to find the original location of the application, and use that
instead of the location it is running from, when looking for the ROM
image and other things. (Mini vMac doesn't look for external library
code, so Path Randomization has no benefit.) This is done using some
undocumented SecTranslocate calls as suggested by "Objective-See" and
Jeff Johnson. Previously, I have recommended using "xattr -rc <path to
Mini vMac>" from the command line to disable Path Randomization. This
will no longer be needed for Mini vMac 36.