Import Mini vMac 180708
The latest Mini vMac Development source snapshot merges in a number of
fixes from Ryan.

The "-min-extn" option wasn't working after the addition of the 'P'
command of the Control Mode (for getting the compile time options).

The source image is now larger so that "-all-src 1" can work.

In the Cocoa version for OS X, the key NSHighResolutionCapable has been
added to the Info.plist file, so that the Mini vMac window frame will
draw in higher resolution on a Retina screen.

There is a new compile time option, "-gse 1", for the Cocoa version for
OS X, which on computers with more than one GPU allows the operating
system to change which one is used. Otherwise on a MacBook pro, it is
reported that the higher power discrete GPU is always used instead of
the lower power integrated GPU, and that besides wasting power for no
noticable benefit, it also takes time to switch on the discrete GPU when
Mini vMac starts. update : I seem to have missed part of this patch.
This will be fixed next snapshot.

"-gse 1" should probably become the default in the future. But I don't
have a MacBook pro to test it on, and I have a few questions about it.

A bug is fixed in the new "-svd 0" option, where in the Cocoa version in
OS X 10.6 and prior, the "out" folder would not be created, if didn't
already exist.