Import Mini vMac 180715
The latest Mini vMac Development source snapshot tries out backtracking
a bit in the build system. Instead of a 68k Macintosh program to
generate files to compile, it uses a cross platform standard C command
line tool to generate a script to write configuration files, as was done
prior to Mini vMac 3 over a decade ago.

This should allow the Variations Service to be more efficient, and is
also more convenient for experienced programmers. The trade off is in
making it harder for others to compile their own variations. Now I would
prefer for people to use the Variations Service, and this is not only
because of the funding model. Helping beginners figure out how to
compile does not scale. And trying to making Mini vMac work well with
random versions of random compilers also does not scale. With the
Variations Service I can ensure quality of the compiled variations. The
Variations Service has greatly improved in past months, fully automatic
with 30 second turnaround (and still getting faster).

The source snapshot is now a standard compressed archive, in ".tgz"
format. The source for the configuration tool is in "setup"folder. In a
unix like system it can be compiled using something like "gcc
setup/tool.c -o setup_t", then run with something like "./setup_t -t
mc64 > setup.sh", and the output run with ". ./setup.sh", and then Mini
vMac can be compiled with "make".

The new configuration tool is still a work in progress. It should be
better documented when it is more settled.