Import Mini vMac 180729
The latest Mini vMac Development source snapshot adds a new way of
finding the ROM image.

It was pointed out to me that using undocumented calls is not a good
idea, so I have disabled the new code for working around the silly Path
Randomization misfeature of OS X, which prevents finding the ROM image
in the folder containing Mini vMac.

Instead, on all platforms (not just OS X), if the ROM image is not found
in any of the locations that Mini vMac looks, the error can now be
recovered from. You can drag the ROM image onto Mini vMac, and booting
will resume. (Any of the methods that would normally mount a disk image
will work to get the ROM, such as the Open dialog.)