Import Mini vMac 180805
The latest Mini vMac Development source snapshot fixes support for all
supported IDEs in the new configuration tool. The new tool always uses
the equivalent of "-cfg 1 -no-src 1" in the old build system, but these
options were previously only implemented for "-e mvc".

Also, Apple Xcode 9.4.1 is now supported.

And also, a memory leak is fixed in the Cocoa port for OS X, when
exporting the clipboard. This was found by the Xcode Analyze feature.