Import Mini vMac 180826
Today's Development source snapshot removes the length limit on the
Control-P command to copy the variations options, and adds an option to
disable code signing.

Previously, converting a string from a cross platform C string constant
to a representation that can be placed on the host clipboard involved a
fixed sized buffer. Now that part of the conversion is done in two
passes, first to find the length of the result, and second to actually
produce the result. (The fixed size was 512 bytes, so it would be
unusual to have hit this limit. It would need quite a lot of variation
options.)

While in this area, I added the "copy variation options" command to the
list in the help screen (Control-H). It would be nice to have
translations of this string, and also the string "Variation options
copied", to the other ten languages supported by Mini vMac 36.

Also while in this area, I noticed that in the help screen, the original
intent appears to have been that the letters on the left would be
surrounded by single quotes. But that displaying a single quote wasn't
implemented properly and showed up as space. After implementing
displaying the single quote, I think it looks better without. However, I
noticed that preceding the "-" there are two spaces, since one was
supposed to be the single quote. Somehow I never noticed this before. I
have changed that to a single space.

Recently code signing has been added to the Cocoa OS X version of Mini
vMac. But it has disadvantages. Apple could choose to revoke the
certificate any time, and Mini vMac would stop working. Also, since Mini
vMac doesn't use Apple's time stamp server (because that would break
reproducible builds), Mini vMac will probably stop working when the
certificate expires, 5 years after being issued. (The plan is to renew
the certificate every year, so that a compile of Mini vMac should work
for at least 4 years.)

Some people may prefer to avoid these issues, so I have added an "-sgn
0" option to disable code signing. The disadvantage is that OS X will by
default refuse to run it (as happens with Mini vMac 3.5). This new
option can be used in the Advanced Variations Service.