Import Mini vMac 180902
Today's Development source snapshot has some assorted fixes in
preparation for moving to beta.

Reconsidering a feedback about the control mode, I changed my mind and
decided that variations of Mini vMac without the control mode should not
be supported. The build system will now consider it an error if "-km"
options result in no key mapping to the control mode.

While in that area, if the "control" key is not mapped to the control
mode, but some other key is, the user interface strings are now
adjusted.

Also, when using the "-km" options, it may not make sense for the K
command of the control mode to toggle the emulated control key. There is
a new option "-ekt <dst>" to choose which emulated key to toggle. The
user interface strings are adjusted accordingly. When this option is not
chosen, Mini vMac will pick an emulated key that is not mapped to any
real key, instead of always the control key.

I have concluded that "Abnormal Situation" reports in the regular non
debug versions of Mini vMac are not on the whole helpful, and so are now
disabled by default. There is a new compile time option "-abr 1" to
enable them.

I spent a few days tracking down a twice reported issue when compiling
with Microsoft Visual Studio 2017. It worked fine with an older version
of Studio 2017 that I had, but once I downloaded the latest version, I
was able to reproduce the problem. Turning off optimization made the
problem go away. This compiler provides a way to set optimization level
per function, so I was able to narrow it down to single simple function,
where it was generating an "shr" instruction when it should have been an
"sar" instruction. I don't see any way this is undefined C behavior, so
it looks like a compiler bug. To maximize the chance of a correct
compile, I have changed the Mini vMac build system to disable
optimization for this compiler. (This compiler is not used for the
versions of Mini vMac that I provide.)

Trying to using the Location and Time Zone options (-alc, -lcy, -lcx,
-atz, -lcd, -lczs, -lcz) when emulating a Mac 128K, 512K, or 512Ke will
now generate an error. These options are saved in the Extended Parameter
RAM, which these machines don't have.