Import Mini vMac 36.02
Today's Mini vMac 36.02 beta has a few fixes.

The game PowerMonger was reported not to work in Mini vMac 3.5.8, though
it worked in 3.4.1. This was due to a bug in emulating the TRAPcc
instruction.

While looking into a report of issues with Num Lock, I found that on
Windows, if Num Lock is on, the operating passes virtual key codes like
VK_NUMPAD0 through VK_NUMPAD9, but if Num Lock is off, it passes virtual
key codes like VK_LEFT, VK_RIGHT, VK_UP, and VK_DOWN. A flag bit
indicates whether VK_LEFT means the left arrow key or the 4 key on the
keypad. So now Mini vMac will check that flag, and map the keys on the
keypad to the emulated Macintosh keypad, regardless of whether Num Lock
is on. (But this doesn't fully explain the reported issue.)

It also turns out that on Windows (at least on some machines and os
versions) the operating system does not use virtual key codes
VK_LCONTROL/VK_RCONTROL/VK_LMENU/VK_RMENU, but instead uses
VK_CONTROL/VK_MENU with that same flag bit to indicate whether it was
the left or right key. If Mini vMac is compiled with "-km" options that
require distinguishing between left and right, then that is checked for
VK_CONTROL/VK_MENU.

Also on Windows, I noticed that the "menu key" (VK_APPS), is usually on
the right side of key (if present at all), and so now maps to the right
option key instead of the left, if -km options require distinguishing
them.

The Cocoa OS X port now implements distinguishing between left and right
modifier keys. But this requires looking at undocumented flags of the
key event, which might or might not always work, so it is not done
unless -km options are used that require it. (I have not found any way
to distinguish left and right modifiers in the Carbon OS X port.)

Also in the Cocoa OS X port, launching Mini vMac by drag and drop of a
disk image on to the application icon would not work because of the new
code to allow drag and drop of the ROM image. Some initialization stuff
for Cocoa caused file open requests to be received earlier than
expected, so that initialization has been moved later, after the ROM has
been looked for in the usual places.