Import Mini vMac 36.04
The latest Mini vMac 36.04 beta fixes a few issues in the classic
Macintosh port. (The fixes were developed inside Mini vMac.)

The new "-svd 0" option is now implemented in classic Macintosh port.
While doing this, it turned out that the build system was generating an
incorrect Rez command for the MPW development environment. It also
turned out that if debug logging was turned on for this port, and the
log already existed when Mini vMac started, the log wasn't cleared.

An unrelated change is that default version of XCode chosen by the build
system is now 9.4.1, the latest version supported. Since XCode is no
longer used for official Mini vMac builds, there is no reason to default
to the specific antique version previously used. Similarly, Microsoft
Visual Studio 2017 is now the default version chosen for that
development environment.